+++ sphinx/source/changelog.rst
@@ -33,7 +33,8 @@ Changelog
   executing code in a 128K memory snapshot
 * Fixed how :ref:`trace.py` handles the value of the SP register in a 128K SNA
   file
-* Fixed how :ref:`trace.py` logs timestamps when an interrupt occurs
+* Fixed how :ref:`tap2sna.py` and :ref:`trace.py` log timestamps when an
+  interrupt occurs
 
 9.0 (2023-11-04)
 ----------------